Complete Get and Posts
The posts and the get is working. The login is working correctly and works appropoiate. The register is still working,
the check role is not working with the post register. This has to do with the decoded token logging differently than
login post does. Still working on this. Otherwise other mvp is done.